refactored some package naming
