+++ app/src/main/java/com/cgavlabs/jeepforecast/events/DataSavedEvent.java
@@ -1,0 +1,4 @@
+package com.cgavlabs.jeepforecast.events;
+
+public class DataSavedEvent {
+}
