+++ app/src/main/java/com/cgavlabs/jeepforecast/main/MainComponent.java
@@ -1,6 +1,7 @@
 package com.cgavlabs.jeepforecast.main;
 
 import com.cgavlabs.jeepforecast.di.AppComponent;
+import com.cgavlabs.jeepforecast.di.UserScope;
 import dagger.Component;
 
 @UserScope @Component(dependencies = { AppComponent.class }, modules = {
