--- app/src/main/java/com/cgavlabs/jeepforecast/models/DataSavedEvent.java
@@ -1,4 +1,0 @@
-package com.cgavlabs.jeepforecast.models;
-
-public class DataSavedEvent {
-}
