+++ app/src/main/java/com/cgavlabs/jeepforecast/settings/weatherconfiglist/WeatherConfigComponent.java
@@ -1,7 +1,7 @@
 package com.cgavlabs.jeepforecast.settings.weatherconfiglist;
 
 import com.cgavlabs.jeepforecast.di.AppComponent;
-import com.cgavlabs.jeepforecast.main.UserScope;
+import com.cgavlabs.jeepforecast.di.UserScope;
 import dagger.Component;
 
 @UserScope @Component(dependencies = { AppComponent.class }, modules = {
